sixth_homework/requirements.txt
tensorflow
Pillow
requests==2.28.1
numpy==1.23.1
argparse